ServiciosBanco/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.10.1
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
openapi-codec==1.3.2
outcome==1.2.0
packaging==23.1
pika==1.3.2
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.21
PyMySQL==1.0.3
PySocks==1.7.1
pytz==2023.3
reportlab==4.0.4
requests==2.29.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.9.0
simplejson==3.19.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wsproto==1.2.0
